fix tn -> tx
